requirements.txt
# CDC PITR System Requirements

# Database connector
psycopg2-binary>=2.9.0

# CLI table formatting
tabulate>=0.9.0

# Optional: For better performance with large JSON files
# orjson>=3.9.0

# Installation:
# pip install -r requirements.txt
